add response logging
