+++ frontend/src/app/api/auth/session/route.ts
@@ -1,19 +1,7 @@
-import { getSession, refreshSession } from "src/services/auth/session";
-import { isExpiring } from "src/utils/dateUtil";
+import { respondWithTraceAndLogs } from "src/utils/apiUtils";
 
-import { NextResponse } from "next/server";
+import { getUserSession } from "./handler";
 
 export const revalidate = 0;
 
-export async function GET() {
-  const currentSession = await getSession();
-  if (currentSession) {
-    if (isExpiring(currentSession.expiresAt)) {
-      const refreshedSession = await refreshSession(currentSession.token);
-      return NextResponse.json(refreshedSession);
-    }
-    return NextResponse.json(currentSession);
-  } else {
-    return NextResponse.json({ token: "" });
-  }
-}
+export const GET = respondWithTraceAndLogs(getUserSession);
